scripts/utils/random_seed_set.py
import random

import numpy as np
import torch


def set_random_seed(seed: int):
    torch.manual_seed(seed)
    torch.cuda.manual_seed_all(seed)
    np.random.seed(seed)
    random.seed(seed)
    torch.set_printoptions(precision=5)
