requirements.txt
einops
jaxtyping
omegaconf
transformers
open_clip_torch
trimesh
numpy
huggingface-hub
rembg[gpu]; sys_platform != 'darwin'
rembg; sys_platform == 'darwin'
git+https://github.com/vork/PyNanoInstantMeshes.git
gpytoolbox
./texture_baker/
./scripts/
wandb
sentencepiece
diffusers
accelerate
jaxlib==0.4.30
./
torchao==0.4.0